+++ app/src/main/java/edu/uco/houselannister/saveasingle/activities/LoginActivity.java
@@ -34,7 +34,7 @@ import java.util.List;
 import edu.uco.houselannister.saveasingle.R;
 import edu.uco.houselannister.saveasingle.domain.Authentication;
 import edu.uco.houselannister.saveasingle.model.AppModel;
-import edu.uco.houselannister.saveasingle.serivce.AppService;
+import edu.uco.houselannister.saveasingle.service.AppService;
 
 import static android.Manifest.permission.READ_CONTACTS;
