+++ app/src/test/java/edu/uco/houselannister/saveasingle/model/AppModelTest.java
@@ -6,7 +6,7 @@ import org.junit.Test;
 
 import edu.uco.houselannister.saveasingle.domain.Model;
 import edu.uco.houselannister.saveasingle.domain.Questionnaire;
-import edu.uco.houselannister.saveasingle.serivce.AppService;
+import edu.uco.houselannister.saveasingle.service.AppService;
 
 import static org.junit.Assert.*;
